LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/going-out-of-flow-after-spanner-expected.txt


FAIL Going out of flow and still fragmented assert_equals: expected 0 but got 100
